Add THREAD_LOCK_DEBUG code
This can be used to help debug weird thread locking behavior that can be hard to debug with a debugger (especially since the thread lock is taken in lots of different places throughout the codebase). Maybe it should be generalized to be in `simple_locks.h` for any kind of simple lock?

I used it to successfully debug a deadlock between `thread_timer_expire` and `waitq_wakeup64_all_locked`. I'm committing this because I think it might help with similar issues in the future. To enable it, just add `-DTHREAD_LOCK_DEBUG=1` to `DEFINES` in the Makefile.