Fix deadlock between `thread_timer_expire` and `waitq_wakeup64_all_locked`
It looks like originally `thread_call_cancel` didn't wait for the work to be canceled, but commit 7b5990dfbdf1df783c4b25c54c99a6c35d99a0f2 changed that to try to fix thread call scheduling. In my testing, reverting this change has not broken anything, so I think this is fine.

This also makes it safe to use now in an IRQ context, so it fixes some crashes that were happening when using kqueue with network sockets.