Revert c971596044c0dc2f21b1c78a8bf315adff070996
I had originally reverted the other commit because I thought it was related to the AppKit issue (as it turned out, it was not). So I'm reverting it back because it seems to not cause any bugs and in fact fixes a "scheduling while atomic" bug I was seeing with rustup that was locking it up randomly.