Workaround: Don't tell mach_vm_remap users if the memory is executable
For some reason, LLDB breaks if we tell it the remapped memory is executable. But we *have* to make it executable if the original memory is executable because other users of mach_vm_remap/vm_remap (like objc-trampolines) expect this behavior. Fortunately, they don't check the result of `current_protection`. So let's just let it be our little secret and not tell callers the memory is executable.