Use TASK_INTERRUPTIBLE in thread_suspended_logic
This is probably wrong, but so was the previous code, because suspended processes are supposed to be able to be woken up with SIGCONT (and the previous code did not allow this). As noted in the comment, the proper behavior would be to use `wake_up_state`, but Linux doesn't export that, unfortunately.

This enables process spawned with POSIX_SPAWN_START_SUSPENDED to be continued with SIGCONT.

During testing, I didn't find any issues with LLDB due to this change.